+++ TeachingRecordSystem/src/TeachingRecordSystem.Api/V3/V20240101/Controllers/TeachersController.cs
@@ -30,8 +30,11 @@ public class TeachersController(IMapper mapper) : ControllerBase
             trn,
             include is not null ? (GetPersonCommandIncludes)include : GetPersonCommandIncludes.None,
             DateOfBirth: null,
-            ApplyLegacyAlertsBehavior: true,
-            ApplyAppropriateBodyUserRestrictions: false);
+            NationalInsuranceNumber: null,
+            new GetPersonCommandOptions()
+            {
+                ApplyLegacyAlertsBehavior = true
+            });
 
         var result = await handler.HandleAsync(command);