+++ TeachingRecordSystem/src/TeachingRecordSystem.Api/V3/V20240920/Controllers/PersonsController.cs
@@ -33,8 +33,11 @@ public class PersonsController(IMapper mapper) : ControllerBase
             trn,
             (GetPersonCommandIncludes)include,
             dateOfBirth,
-            ApplyLegacyAlertsBehavior: false,
-            ApplyAppropriateBodyUserRestrictions: User.IsInRole(ApiRoles.AppropriateBody));
+            NationalInsuranceNumber: null,
+            new GetPersonCommandOptions()
+            {
+                ApplyAppropriateBodyUserRestrictions = User.IsInRole(ApiRoles.AppropriateBody)
+            });
 
         var result = await handler.HandleAsync(command);