+++ TeachingRecordSystem/src/TeachingRecordSystem.Api/V3/V20250203/Responses/FindPersonResponse.cs
@@ -1,3 +1,4 @@
+using AutoMapper.Configuration.Annotations;
 using TeachingRecordSystem.Api.V3.Implementation.Operations;
 using TeachingRecordSystem.Api.V3.V20250203.Requests;
 using TeachingRecordSystem.Core.ApiSchema.V3.V20240101.Dtos;
@@ -27,6 +28,7 @@ public record FindPersonResponseResult
     public required QtsInfo? Qts { get; init; }
     public required EytsInfo? Eyts { get; init; }
     public required IReadOnlyCollection<Alert> Alerts { get; init; }
+    [SourceMember("Induction.Status")]
     public required InductionStatus InductionStatus { get; init; }
     public required QtlsStatus QtlsStatus { get; set; }
 }