+++ TeachingRecordSystem/src/TeachingRecordSystem.Api/V3/V20250327/Controllers/PersonController.cs
@@ -29,8 +29,7 @@ public class PersonController(IMapper mapper) : ControllerBase
             Trn: User.FindFirstValue("trn")!,
             include is not null ? (GetPersonCommandIncludes)include : GetPersonCommandIncludes.None,
             DateOfBirth: null,
-            ApplyLegacyAlertsBehavior: false,
-            ApplyAppropriateBodyUserRestrictions: false);
+            NationalInsuranceNumber: null);
 
         var result = await handler.HandleAsync(command);