+++ TeachingRecordSystem/src/TeachingRecordSystem.Api/V3/V20250327/Responses/FindPersonsResponse.cs
@@ -29,6 +29,7 @@ public record FindPersonsResponseResult
     public required QtsInfo? Qts { get; init; }
     public required EytsInfo? Eyts { get; init; }
     public required IReadOnlyCollection<Alert> Alerts { get; init; }
+    [SourceMember("Induction.Status")]
     public required InductionStatus InductionStatus { get; init; }
     public required QtlsStatus QtlsStatus { get; set; }
 }