+++ TeachingRecordSystem/src/TeachingRecordSystem.Api/V3/VNext/Controllers/PersonController.cs
@@ -1,3 +1,4 @@
+using System.Security.Claims;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Swashbuckle.AspNetCore.Annotations;
@@ -10,7 +11,7 @@ using TeachingRecordSystem.Api.V3.VNext.Responses;
 namespace TeachingRecordSystem.Api.V3.VNext.Controllers;
 
 [Route("person")]
-public class PersonController : ControllerBase
+public class PersonController(IMapper mapper) : ControllerBase
 {
     [Authorize(AuthorizationPolicies.IdentityUserWithTrn)]
     [HttpGet]
@@ -20,10 +21,19 @@ public class PersonController : ControllerBase
         Description = "Gets the details for the authenticated person.")]
     [ProducesResponseType(typeof(GetPersonResponse), StatusCodes.Status200OK)]
     [ProducesResponseType(typeof(void), StatusCodes.Status403Forbidden)]
-    public Task<IActionResult> GetAsync(
+    public async Task<IActionResult> GetAsync(
         [FromQuery, ModelBinder(typeof(FlagsEnumStringListModelBinder)), SwaggerParameter("The additional properties to include in the response.")] GetPersonRequestIncludes? include,
         [FromServices] GetPersonHandler handler)
     {
-        throw new NotImplementedException();
+        var command = new GetPersonCommand(
+            Trn: User.FindFirstValue("trn")!,
+            include is not null ? (GetPersonCommandIncludes)include : GetPersonCommandIncludes.None,
+            DateOfBirth: null,
+            NationalInsuranceNumber: null);
+
+        var result = await handler.HandleAsync(command);
+
+        return result.ToActionResult(r => Ok(mapper.Map<GetPersonResponse>(r)))
+            .MapErrorCode(ApiError.ErrorCodes.PersonNotFound, StatusCodes.Status403Forbidden);
     }
 }