+++ TeachingRecordSystem/src/TeachingRecordSystem.Api/V3/VNext/Responses/GetPersonResponse.cs
@@ -1,10 +1,12 @@
+using OneOf;
 using Optional;
 using TeachingRecordSystem.Api.V3.Implementation.Operations;
 using TeachingRecordSystem.Core.ApiSchema.V3.VNext.Dtos;
 using Alert = TeachingRecordSystem.Core.ApiSchema.V3.V20240920.Dtos.Alert;
+using EytsInfo = TeachingRecordSystem.Core.ApiSchema.V3.VNext.Dtos.EytsInfo;
 using InductionInfo = TeachingRecordSystem.Core.ApiSchema.V3.VNext.Dtos.InductionInfo;
 using NameInfo = TeachingRecordSystem.Core.ApiSchema.V3.V20240101.Dtos.NameInfo;
-using ProfessionalStatusStatus = TeachingRecordSystem.Core.ApiSchema.V3.V20250425.Dtos.ProfessionalStatusStatus;
+using ProfessionalStatusStatus = TeachingRecordSystem.Core.ApiSchema.V3.VNext.Dtos.ProfessionalStatusStatus;
 using QtlsStatus = TeachingRecordSystem.Core.ApiSchema.V3.V20250203.Dtos.QtlsStatus;
 using QtsInfo = TeachingRecordSystem.Core.ApiSchema.V3.VNext.Dtos.QtsInfo;
 
@@ -24,13 +26,13 @@ public record GetPersonResponse
     public required string? EmailAddress { get; init; }
     public required QtsInfo? Qts { get; init; }
     public required EytsInfo? Eyts { get; init; }
-    public required InductionInfo Induction { get; init; }
-    public required Option<IReadOnlyCollection<GetPersonResponseRouteToProfessionalStatus>> RoutesToProfessionalStatuses { get; init; }
+    public required Option<InductionInfo> Induction { get; init; }
+    public required Option<OneOf<IReadOnlyCollection<GetPersonResponseRouteToProfessionalStatus>, IReadOnlyCollection<GetPersonResponseRouteToProfessionalStatusForAppropriateBody>>> RoutesToProfessionalStatuses { get; init; }
     public required Option<IReadOnlyCollection<GetPersonResponseMandatoryQualification>> MandatoryQualifications { get; init; }
     public required Option<IReadOnlyCollection<Alert>> Alerts { get; init; }
     public required Option<IReadOnlyCollection<NameInfo>> PreviousNames { get; init; }
     public required Option<bool> AllowIdSignInWithProhibitions { get; init; }
-    public required QtlsStatus QtlsStatus { get; set; }
+    public required QtlsStatus QtlsStatus { get; init; }
 }
 
 [AutoMap(typeof(GetPersonResultMandatoryQualification))]
@@ -41,12 +43,13 @@ public record GetPersonResponseMandatoryQualification
     public required string Specialism { get; init; }
 }
 
+[AutoMap(typeof(GetPersonResultRouteToProfessionalStatus))]
 public record GetPersonResponseRouteToProfessionalStatus
 {
     public required Guid RouteToProfessionalStatusId { get; init; }
     public required RouteToProfessionalStatusType RouteToProfessionalStatusType { get; init; }
     public required ProfessionalStatusStatus Status { get; init; }
-    public required DateOnly? AwardedDate { get; init; }
+    public required DateOnly? HoldsFrom { get; init; }
     public required DateOnly? TrainingStartDate { get; init; }
     public required DateOnly? TrainingEndDate { get; init; }
     public required IReadOnlyCollection<TrainingSubject> TrainingSubjects { get; init; }
@@ -54,5 +57,18 @@ public record GetPersonResponseRouteToProfessionalStatus
     public required TrainingCountry? TrainingCountry { get; init; }
     public required TrainingProvider? TrainingProvider { get; init; }
     public required DegreeType? DegreeType { get; init; }
-    public required InductionExemptionReason? InductionExemption { get; set; }
+    public required GetPersonResponseProfessionalStatusInductionExemption InductionExemption { get; init; }
+}
+
+[AutoMap(typeof(GetPersonResultRouteToProfessionalStatusForAppropriateBody))]
+public record GetPersonResponseRouteToProfessionalStatusForAppropriateBody
+{
+    public required TrainingProvider TrainingProvider { get; init; }
+}
+
+[AutoMap(typeof(GetPersonResultRouteToProfessionalStatusInductionExemption))]
+public record GetPersonResponseProfessionalStatusInductionExemption
+{
+    public required bool IsExempt { get; init; }
+    public required IReadOnlyCollection<InductionExemptionReason> ExemptionReasons { get; init; }
 }